requirements.txt
# 现有依赖
requests>=2.28.0
openpyxl>=3.0.0

# ML新增依赖
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pyarrow>=14.0.0  # Parquet支持
python-dateutil>=2.8.0
pyyaml>=6.0  # YAML配置支持